SFTP-File-Routing/requirements.txt
Cryptography
Paramiko
python-gnupg
pysftp